requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
blinker>=1.7.0
email-validator>=2.1.0
stripe>=7.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
itsdangerous>=2.1.2
WTForms>=3.1.1
Werkzeug>=3.0.0
qrcode>=7.4.2
gunicorn>=21.2.0
requests>=2.31.0
